fix(hashtag): #122 해시태그 삭제 함수화 및 삭제 순서 변경

+++ backend/src/main/java/com/bremen/backend/domain/article/service/ArticleHashtagService.java
@@ -8,5 +8,5 @@ import com.bremen.backend.domain.article.entity.Article;
 public interface ArticleHashtagService {
 	List<String> addHashtags(Article article, Set<String> hashtags);
 
-	Long removeHashtags(Long articleId);
+	int removeHashtags(Long articleId);
 }
